+++ cmail-smtpd/mail.c
@@ -122,9 +122,9 @@ int mail_line(LOGGER log, MAIL* mail, char* line){
 int mail_recvheader(LOGGER log, MAIL* mail, char* announce, bool suppress_submitter){
 	char time_buffer[SMTP_HEADER_LINE_MAX];
 	char* recv_header = NULL;
-	unsigned header_allocated = 0;
+	size_t header_allocated = 0;
 
-	unsigned mark=0, i, off=0;
+	unsigned mark = 0, i, off = 0;
 
 	MAILPATH* forward_path = (mail->forward_paths[1]) ? NULL:mail->forward_paths[0];
