+++ lib/common.c
@@ -69,7 +69,7 @@ int common_strrepl(char* buffer, unsigned length, char* variable, char* replacem
 	return 0;
 }
 
-char* common_strappf(char* target, unsigned* target_allocated, char* fmt, ...){
+char* common_strappf(char* target, size_t* target_allocated, char* fmt, ...){
 	va_list args, copy;
 
 	size_t target_len = target ? strlen(target):0;
@@ -170,12 +170,12 @@ ssize_t common_next_line(LOGGER log, char* buffer, size_t* append_offset_p, ssiz
 
 	logprintf(log, LOG_DEBUG, "Next line parser called with offset %d bytes %d\n", append_offset, new_bytes);
 
-	if(append_offset > 1 && buffer[append_offset-1] == 0 && buffer[append_offset-2] == 0){
+	if(append_offset > 1 && buffer[append_offset - 1] == 0 && buffer[append_offset - 2] == 0){
 		//copyback clearing of last line
 
-		for(i=0;i<new_bytes;i++){
+		for(i = 0; i < new_bytes; i++){
 			//logprintf(log, LOG_DEBUG, "Moving character %02X from position %d to %d\n", client_data->recv_buffer[client_data->recv_offset+i+1+c], client_data->recv_offset+i+1+c, c);
-			buffer[i] = buffer[append_offset+i];
+			buffer[i] = buffer[append_offset + i];
 		}
 
 		append_offset = 0;
@@ -184,7 +184,7 @@ ssize_t common_next_line(LOGGER log, char* buffer, size_t* append_offset_p, ssiz
 	}
 
 	//scan new bytes for terminators
-	for(i=0;i<new_bytes-1;i++){ //last byte is checked in condition
+	for(i = 0; i < new_bytes - 1; i++){ //last byte is checked in condition
 		if(buffer[append_offset + i] == '\r' && buffer[append_offset + i + 1] == '\n'){
 			//terminate line
 			buffer[append_offset + i] = 0;
